requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
crefi==2.0.9
Flask==1.1.2
future==0.18.2
htmlmin==0.1.12
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
logger==1.4
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pymdown-extensions==7.1
python-dotenv==0.14.0
python-twitter==3.5
pytz==2020.1
pyxattr==0.7.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
webassets==2.0
Werkzeug==1.0.1
